Move functions from `helpers.py` to `committee_helpers.py` and `epoch_processing_helpers.py`
